bebob: split into bebob crate
The protocol to operate ASICs of BridgeCo Enhanced Break Out Box and
vendor-specific models is implemented just for a service program,
however it's convenient to have a crate for it.

This commit adds a crate as 'bebob' and push the implementation into it.
