+++ libs/bebob/Cargo.toml
@@ -1,9 +1,13 @@
 [package]
-name = "snd_fw_ctl_services"
+name = "bebob"
 version = "0.1.0"
 authors = ["Takashi Sakamoto <o-takashi@sakamocchi.jp>"]
 edition = "2018"
-license-file = "COPYING"
+license = "GPL-3.0-or-later"
+publish = false
+description = """
+Implementation of protocol for ASIC in BridgeCo Enhanced Break Out Box and vendor-specific models
+"""
 
 [dependencies]
 libc = "0.2"
@@ -11,7 +15,5 @@ nix = "0.17"
 glib = "0.10"
 hinawa = { git = "https://github.com/alsa-project/hinawa-rs.git", tag = "v0.3.0", version = "0.3" }
 alsactl = { git = "https://github.com/alsa-project/alsa-gobject-rs.git", tag = "v0.1.1", version = "0.1" }
-alsaseq = { git = "https://github.com/alsa-project/alsa-gobject-rs.git", tag = "v0.1.1", version = "0.1" }
-core = { path = "core" }
-ieee1212 = { path = "ieee1212" }
-ta1394 = { path = "ta1394" }
+core = { path = "../core" }
+ta1394 = { path = "../ta1394" }
