+++ libs/bebob/src/maudio/ozonic_model.rs
@@ -4,43 +4,38 @@ use glib::Error;
 
 use hinawa::{FwFcpExt, SndUnitExt};
 
-use alsactl::{ElemValueExt, ElemValueExtManual};
-
 use core::card_cntr;
 use card_cntr::{CtlModel, MeasureModel};
 
 use ta1394::{AvcAddr, MUSIC_SUBUNIT_0, Ta1394Avc};
 use ta1394::general::UnitInfo;
-use ta1394::ccm::{SignalAddr, SignalSubunitAddr, SignalUnitAddr};
-use ta1394::audio::{AUDIO_SUBUNIT_0_ADDR, CtlAttr, AudioSelector};
+use ta1394::ccm::{SignalAddr, SignalSubunitAddr};
 
-use crate::bebob::common_ctls::ClkCtl;
-
-use crate::bebob::BebobAvc;
+use super::super::BebobAvc;
+use super::super::common_ctls::ClkCtl;
 
 use super::common_proto::FCP_TIMEOUT_MS;
 use super::normal_ctls::{MeterCtl, MixerCtl, InputCtl};
 
-pub struct SoloModel<'a>{
+pub struct OzonicModel<'a>{
     avc: BebobAvc,
-    clk_ctl: ClkCtl<'a>,
     req: hinawa::FwReq,
+    clk_ctl: ClkCtl<'a>,
     meter_ctl: MeterCtl<'a>,
     mixer_ctl: MixerCtl<'a>,
     input_ctl: InputCtl<'a>,
 }
 
-impl<'a> SoloModel<'a> {
+impl<'a> OzonicModel<'a> {
     const CLK_DST: SignalAddr = SignalAddr::Subunit(SignalSubunitAddr{
         subunit: MUSIC_SUBUNIT_0,
-        plug_id: 0x01,
+        plug_id: 0x05,
     });
     const CLK_SRCS: &'a [SignalAddr] = &[
         SignalAddr::Subunit(SignalSubunitAddr{
             subunit: MUSIC_SUBUNIT_0,
-            plug_id: 0x01,
+            plug_id: 0x05,
         }),
-        SignalAddr::Unit(SignalUnitAddr::Ext(0x01)),
     ];
     const CLK_LABELS: &'a [&'a str] = &["Internal", "S/PDIF"];
 
@@ -56,23 +51,23 @@ impl<'a> SoloModel<'a> {
         "stream-in-1", "stream-in-2", "stream-in-3", "stream-in-4",
     ];
 
-    const MIXER_DST_FB_IDS: &'a [u8] = &[0x01, 0x01];
+    const MIXER_DST_FB_IDS: &'a [u8] = &[0x01, 0x02];
     const MIXER_LABELS: &'a [&'a str] = &["mixer-1/2", "mixer-3/4"];
-    const MIXER_PHYS_SRC_FB_IDS: &'a [u8] = &[0x00, 0x01];
-    const PHYS_IN_LABELS: &'a [&'a str] = &["analog-1/2", "digital-1/2"];
-    const MIXER_STREAM_SRC_FB_IDS: &'a [u8] = &[0x02, 0x03];
-    const STREAM_IN_LABELS: &'a [&'a str] = &["stream-1/2", "stream-1/2"];
+    const MIXER_PHYS_SRC_FB_IDS: &'a [u8] = &[0x02, 0x03];
+    const PHYS_IN_LABELS: &'a [&'a str] = &["analog-1/2", "analog-3/4"];
+    const MIXER_STREAM_SRC_FB_IDS: &'a [u8] = &[0x00, 0x01];
+    const STREAM_IN_LABELS: &'a [&'a str] = &["stream-1/2", "stream-3/4"];
 
-    const PHYS_IN_FB_IDS: &'a [u8] = &[0x01, 0x02];
-    const STREAM_IN_FB_IDS: &'a [u8] = &[0x03, 0x04];
+    const PHYS_IN_FB_IDS: &'a [u8] = &[0x03, 0x04];
+    const STREAM_IN_FB_IDS: &'a [u8] = &[0x01, 0x02];
 
     pub fn new() -> Self {
-        SoloModel{
+        OzonicModel{
             avc: BebobAvc::new(),
             req: hinawa::FwReq::new(),
             clk_ctl: ClkCtl::new(&Self::CLK_DST, Self::CLK_SRCS, Self::CLK_LABELS),
             meter_ctl: MeterCtl::new(Self::IN_METER_LABELS, Self::STREAM_METER_LABELS, Self::OUT_METER_LABELS,
-                                     false, 0, true),
+                                     false, 0, false),
             mixer_ctl: MixerCtl::new(
                 Self::MIXER_DST_FB_IDS, Self::MIXER_LABELS,
                 Self::MIXER_PHYS_SRC_FB_IDS, Self::PHYS_IN_LABELS,
@@ -86,7 +81,7 @@ impl<'a> SoloModel<'a> {
     }
 }
 
-impl<'a> CtlModel<hinawa::SndUnit> for SoloModel<'a> {
+impl<'a> CtlModel<hinawa::SndUnit> for OzonicModel<'a> {
     fn load(&mut self, unit: &hinawa::SndUnit, card_cntr: &mut card_cntr::CardCntr) -> Result<(), Error> {
         self.avc.fcp.bind(&unit.get_node())?;
 
@@ -98,8 +93,6 @@ impl<'a> CtlModel<hinawa::SndUnit> for SoloModel<'a> {
         self.meter_ctl.load(unit, &self.avc, &self.req, card_cntr)?;
         self.mixer_ctl.load(&self.avc, card_cntr)?;
         self.input_ctl.load(&self.avc, card_cntr)?;
-
-        SpdifOutCtl::load(&self.avc, card_cntr)?;
 
         Ok(())
     }
@@ -114,8 +107,6 @@ impl<'a> CtlModel<hinawa::SndUnit> for SoloModel<'a> {
         } else if self.mixer_ctl.read(&self.avc, elem_id, elem_value)? {
             Ok(true)
         } else if self.input_ctl.read(&self.avc, elem_id, elem_value)? {
-            Ok(true)
-        } else if SpdifOutCtl::read(&self.avc, elem_id, elem_value)? {
             Ok(true)
         } else {
             Ok(false)
@@ -134,15 +125,13 @@ impl<'a> CtlModel<hinawa::SndUnit> for SoloModel<'a> {
             Ok(true)
         } else if self.input_ctl.write(&self.avc, elem_id, old, new)? {
             Ok(true)
-        } else if SpdifOutCtl::write(&self.avc, elem_id, old, new)? {
-            Ok(true)
         } else {
             Ok(false)
         }
     }
 }
 
-impl<'a> MeasureModel<hinawa::SndUnit> for SoloModel<'a> {
+impl<'a> MeasureModel<hinawa::SndUnit> for OzonicModel<'a> {
     fn get_measure_elem_list(&mut self, elem_id_list: &mut Vec<alsactl::ElemId>) {
         elem_id_list.extend_from_slice(&self.meter_ctl.measure_elems);
     }
@@ -158,7 +147,7 @@ impl<'a> MeasureModel<hinawa::SndUnit> for SoloModel<'a> {
     }
 }
 
-impl<'a> card_cntr::NotifyModel<hinawa::SndUnit, bool> for SoloModel<'a> {
+impl<'a> card_cntr::NotifyModel<hinawa::SndUnit, bool> for OzonicModel<'a> {
     fn get_notified_elem_list(&mut self, elem_id_list: &mut Vec<alsactl::ElemId>) {
         elem_id_list.extend_from_slice(&self.clk_ctl.notified_elem_list);
     }
@@ -174,47 +163,3 @@ impl<'a> card_cntr::NotifyModel<hinawa::SndUnit, bool> for SoloModel<'a> {
         self.clk_ctl.read(&self.avc, elem_id, elem_value, FCP_TIMEOUT_MS)
     }
 }
-
-const SPDIF_OUT_SRC_NAME: &str = "S/PDIF-out-source";
-const SPDIF_OUT_SRC_LABELS: &[&str] = &["stream-3/4", "mixer-3/4"];
-const SPDIF_OUT_SRC_FB_ID: u8 = 0x01;
-
-trait SpdifOutCtl : Ta1394Avc {
-    fn load(&self, card_cntr: &mut card_cntr::CardCntr) -> Result<(), Error> {
-        let elem_id = alsactl::ElemId::new_by_name(alsactl::ElemIfaceType::Mixer, 0, 0, SPDIF_OUT_SRC_NAME, 0);
-        let _ = card_cntr.add_enum_elems(&elem_id, 1, 1, SPDIF_OUT_SRC_LABELS, None, true)?;
-
-        Ok(())
-    }
-
-    fn read(&self, elem_id: &alsactl::ElemId, elem_value: &mut alsactl::ElemValue)
-        -> Result<bool, Error>
-    {
-        match elem_id.get_name().as_str() {
-            SPDIF_OUT_SRC_NAME => {
-                let mut op = AudioSelector::new(SPDIF_OUT_SRC_FB_ID, CtlAttr::Current, 0xff);
-                self.status(&AUDIO_SUBUNIT_0_ADDR, &mut op, FCP_TIMEOUT_MS)?;
-                elem_value.set_enum(&[op.input_plug_id as u32]);
-                Ok(true)
-            }
-            _ => Ok(false),
-        }
-    }
-
-    fn write(&self, elem_id: &alsactl::ElemId, _: &alsactl::ElemValue, new: &alsactl::ElemValue)
-        -> Result<bool, Error>
-    {
-        match elem_id.get_name().as_str() {
-            SPDIF_OUT_SRC_NAME => {
-                let mut vals = [0];
-                new.get_enum(&mut vals);
-                let mut op = AudioSelector::new(SPDIF_OUT_SRC_FB_ID, CtlAttr::Current, vals[0] as u8);
-                self.control(&AUDIO_SUBUNIT_0_ADDR, &mut op, FCP_TIMEOUT_MS)?;
-                Ok(true)
-            }
-            _ => Ok(false),
-        }
-    }
-}
-
-impl SpdifOutCtl for BebobAvc {}
