+++ libs/bebob/src/maudio/solo_model.rs
@@ -4,33 +4,33 @@ use glib::Error;
 
 use hinawa::{FwFcpExt, SndUnitExt};
 
+use alsactl::{ElemValueExt, ElemValueExtManual};
+
 use core::card_cntr;
 use card_cntr::{CtlModel, MeasureModel};
 
 use ta1394::{AvcAddr, MUSIC_SUBUNIT_0, Ta1394Avc};
 use ta1394::general::UnitInfo;
 use ta1394::ccm::{SignalAddr, SignalSubunitAddr, SignalUnitAddr};
+use ta1394::audio::{AUDIO_SUBUNIT_0_ADDR, CtlAttr, AudioSelector};
 
-use crate::bebob::common_ctls::ClkCtl;
+use super::super::common_ctls::ClkCtl;
 
-use crate::bebob::BebobAvc;
+use super::super::BebobAvc;
 
 use super::common_proto::FCP_TIMEOUT_MS;
-use super::normal_ctls::{MeterCtl, MixerCtl, InputCtl, AuxCtl, OutputCtl, HpCtl};
+use super::normal_ctls::{MeterCtl, MixerCtl, InputCtl};
 
-pub struct AudiophileModel<'a>{
+pub struct SoloModel<'a>{
     avc: BebobAvc,
-    req: hinawa::FwReq,
     clk_ctl: ClkCtl<'a>,
+    req: hinawa::FwReq,
     meter_ctl: MeterCtl<'a>,
     mixer_ctl: MixerCtl<'a>,
     input_ctl: InputCtl<'a>,
-    aux_ctl: AuxCtl<'a>,
-    output_ctl: OutputCtl<'a>,
-    hp_ctl: HpCtl<'a>,
 }
 
-impl<'a> AudiophileModel<'a> {
+impl<'a> SoloModel<'a> {
     const CLK_DST: SignalAddr = SignalAddr::Subunit(SignalSubunitAddr{
         subunit: MUSIC_SUBUNIT_0,
         plug_id: 0x01,
@@ -40,7 +40,7 @@ impl<'a> AudiophileModel<'a> {
             subunit: MUSIC_SUBUNIT_0,
             plug_id: 0x01,
         }),
-        SignalAddr::Unit(SignalUnitAddr::Ext(0x02)),
+        SignalAddr::Unit(SignalUnitAddr::Ext(0x01)),
     ];
     const CLK_LABELS: &'a [&'a str] = &["Internal", "S/PDIF"];
 
@@ -49,38 +49,30 @@ impl<'a> AudiophileModel<'a> {
     ];
 
     const OUT_METER_LABELS: &'a [&'a str] = &[
-        "analog-out-1", "analog-out-2", "analog-out-3", "analog-out-4",
-        "digital-out-1", "digital-out-2",
+        "analog-out-1", "analog-out-2", "digital-out-1", "digital-out-2",
     ];
 
-    const MIXER_DST_FB_IDS: &'a [u8] = &[0x01, 0x02, 0x03];
-    const MIXER_LABELS: &'a [&'a str] = &["mixer-1/2", "mixer-3/4", "mixer-5/6"];
-    const MIXER_PHYS_SRC_FB_IDS: &'a [u8] = &[0x03, 0x04];
+    const STREAM_METER_LABELS: &'a [&'a str] = &[
+        "stream-in-1", "stream-in-2", "stream-in-3", "stream-in-4",
+    ];
+
+    const MIXER_DST_FB_IDS: &'a [u8] = &[0x01, 0x01];
+    const MIXER_LABELS: &'a [&'a str] = &["mixer-1/2", "mixer-3/4"];
+    const MIXER_PHYS_SRC_FB_IDS: &'a [u8] = &[0x00, 0x01];
     const PHYS_IN_LABELS: &'a [&'a str] = &["analog-1/2", "digital-1/2"];
-    const MIXER_STREAM_SRC_FB_IDS: &'a [u8] = &[0x00, 0x01, 0x02];
-    const STREAM_IN_LABELS: &'a [&'a str] = &["stream-1/2", "stream-3/4", "stream-5/6"];
-    const HP_SRC_LABELS: &'a [&'a str] = &["mixer-1/2", "mixer-3/4", "mixer-5/6", "aux-1/2"];
+    const MIXER_STREAM_SRC_FB_IDS: &'a [u8] = &[0x02, 0x03];
+    const STREAM_IN_LABELS: &'a [&'a str] = &["stream-1/2", "stream-1/2"];
 
-    const PHYS_IN_FB_IDS: &'a [u8] = &[0x04, 0x05];
-    const STREAM_IN_FB_IDS: &'a [u8] = &[0x01, 0x02, 0x03];
-
-    const AUX_OUT_FB_ID: u8 = 0x0b;
-    const AUX_PHYS_SRC_FB_IDS: &'a [u8] = &[0x09, 0x0a];
-    const AUX_STREAM_SRC_FB_IDS: &'a [u8] = &[0x06, 0x07, 0x08];
-
-    const PHYS_OUT_LABELS: &'a [&'a str] = &["analog-1/2", "analog-3/4", "digital-1/2"];
-    const PHYS_OUT_FB_IDS: &'a [u8] = &[0x0c, 0x0d, 0x0e];
-    const PHYS_OUT_SRC_FB_IDS: &'a [u8] = &[0x01, 0x02, 0x03];
-
-    const HP_SRC_FB_ID: u8 = 0x04;
-    const HP_OUT_FB_ID: u8 = 0x0f;
+    const PHYS_IN_FB_IDS: &'a [u8] = &[0x01, 0x02];
+    const STREAM_IN_FB_IDS: &'a [u8] = &[0x03, 0x04];
 
     pub fn new() -> Self {
-        AudiophileModel{
+        SoloModel{
             avc: BebobAvc::new(),
             req: hinawa::FwReq::new(),
             clk_ctl: ClkCtl::new(&Self::CLK_DST, Self::CLK_SRCS, Self::CLK_LABELS),
-            meter_ctl: MeterCtl::new(Self::IN_METER_LABELS, &[], Self::OUT_METER_LABELS, true, 2, true),
+            meter_ctl: MeterCtl::new(Self::IN_METER_LABELS, Self::STREAM_METER_LABELS, Self::OUT_METER_LABELS,
+                                     false, 0, true),
             mixer_ctl: MixerCtl::new(
                 Self::MIXER_DST_FB_IDS, Self::MIXER_LABELS,
                 Self::MIXER_PHYS_SRC_FB_IDS, Self::PHYS_IN_LABELS,
@@ -90,21 +82,11 @@ impl<'a> AudiophileModel<'a> {
                 Self::PHYS_IN_FB_IDS, Self::PHYS_IN_LABELS,
                 Self::STREAM_IN_FB_IDS, Self::STREAM_IN_LABELS,
             ),
-            aux_ctl: AuxCtl::new(Self::AUX_OUT_FB_ID,
-                Self::AUX_PHYS_SRC_FB_IDS, Self::PHYS_IN_LABELS,
-                Self::AUX_STREAM_SRC_FB_IDS, Self::STREAM_IN_LABELS,
-            ),
-            output_ctl: OutputCtl::new(
-                Self::PHYS_OUT_LABELS,
-                Self::PHYS_OUT_FB_IDS,
-                Self::PHYS_OUT_SRC_FB_IDS,
-            ),
-            hp_ctl: HpCtl::new(Self::HP_OUT_FB_ID, Self::HP_SRC_FB_ID, Self::HP_SRC_LABELS),
         }
     }
 }
 
-impl<'a> CtlModel<hinawa::SndUnit> for AudiophileModel<'a> {
+impl<'a> CtlModel<hinawa::SndUnit> for SoloModel<'a> {
     fn load(&mut self, unit: &hinawa::SndUnit, card_cntr: &mut card_cntr::CardCntr) -> Result<(), Error> {
         self.avc.fcp.bind(&unit.get_node())?;
 
@@ -116,9 +98,8 @@ impl<'a> CtlModel<hinawa::SndUnit> for AudiophileModel<'a> {
         self.meter_ctl.load(unit, &self.avc, &self.req, card_cntr)?;
         self.mixer_ctl.load(&self.avc, card_cntr)?;
         self.input_ctl.load(&self.avc, card_cntr)?;
-        self.aux_ctl.load(&self.avc, card_cntr)?;
-        self.output_ctl.load(&self.avc, card_cntr)?;
-        self.hp_ctl.load(&self.avc, card_cntr)?;
+
+        SpdifOutCtl::load(&self.avc, card_cntr)?;
 
         Ok(())
     }
@@ -134,11 +115,7 @@ impl<'a> CtlModel<hinawa::SndUnit> for AudiophileModel<'a> {
             Ok(true)
         } else if self.input_ctl.read(&self.avc, elem_id, elem_value)? {
             Ok(true)
-        } else if self.aux_ctl.read(&self.avc, elem_id, elem_value)? {
-            Ok(true)
-        } else if self.output_ctl.read(&self.avc, elem_id, elem_value)? {
-            Ok(true)
-        } else if self.hp_ctl.read(&self.avc, elem_id, elem_value)? {
+        } else if SpdifOutCtl::read(&self.avc, elem_id, elem_value)? {
             Ok(true)
         } else {
             Ok(false)
@@ -157,11 +134,7 @@ impl<'a> CtlModel<hinawa::SndUnit> for AudiophileModel<'a> {
             Ok(true)
         } else if self.input_ctl.write(&self.avc, elem_id, old, new)? {
             Ok(true)
-        } else if self.aux_ctl.write(&self.avc, elem_id, old, new)? {
-            Ok(true)
-        } else if self.output_ctl.write(&self.avc, elem_id, old, new)? {
-            Ok(true)
-        } else if self.hp_ctl.write(&self.avc, elem_id, old, new)? {
+        } else if SpdifOutCtl::write(&self.avc, elem_id, old, new)? {
             Ok(true)
         } else {
             Ok(false)
@@ -169,7 +142,7 @@ impl<'a> CtlModel<hinawa::SndUnit> for AudiophileModel<'a> {
     }
 }
 
-impl<'a> MeasureModel<hinawa::SndUnit> for AudiophileModel<'a> {
+impl<'a> MeasureModel<hinawa::SndUnit> for SoloModel<'a> {
     fn get_measure_elem_list(&mut self, elem_id_list: &mut Vec<alsactl::ElemId>) {
         elem_id_list.extend_from_slice(&self.meter_ctl.measure_elems);
     }
@@ -185,7 +158,7 @@ impl<'a> MeasureModel<hinawa::SndUnit> for AudiophileModel<'a> {
     }
 }
 
-impl<'a> card_cntr::NotifyModel<hinawa::SndUnit, bool> for AudiophileModel<'a> {
+impl<'a> card_cntr::NotifyModel<hinawa::SndUnit, bool> for SoloModel<'a> {
     fn get_notified_elem_list(&mut self, elem_id_list: &mut Vec<alsactl::ElemId>) {
         elem_id_list.extend_from_slice(&self.clk_ctl.notified_elem_list);
     }
@@ -201,3 +174,47 @@ impl<'a> card_cntr::NotifyModel<hinawa::SndUnit, bool> for AudiophileModel<'a> {
         self.clk_ctl.read(&self.avc, elem_id, elem_value, FCP_TIMEOUT_MS)
     }
 }
+
+const SPDIF_OUT_SRC_NAME: &str = "S/PDIF-out-source";
+const SPDIF_OUT_SRC_LABELS: &[&str] = &["stream-3/4", "mixer-3/4"];
+const SPDIF_OUT_SRC_FB_ID: u8 = 0x01;
+
+trait SpdifOutCtl : Ta1394Avc {
+    fn load(&self, card_cntr: &mut card_cntr::CardCntr) -> Result<(), Error> {
+        let elem_id = alsactl::ElemId::new_by_name(alsactl::ElemIfaceType::Mixer, 0, 0, SPDIF_OUT_SRC_NAME, 0);
+        let _ = card_cntr.add_enum_elems(&elem_id, 1, 1, SPDIF_OUT_SRC_LABELS, None, true)?;
+
+        Ok(())
+    }
+
+    fn read(&self, elem_id: &alsactl::ElemId, elem_value: &mut alsactl::ElemValue)
+        -> Result<bool, Error>
+    {
+        match elem_id.get_name().as_str() {
+            SPDIF_OUT_SRC_NAME => {
+                let mut op = AudioSelector::new(SPDIF_OUT_SRC_FB_ID, CtlAttr::Current, 0xff);
+                self.status(&AUDIO_SUBUNIT_0_ADDR, &mut op, FCP_TIMEOUT_MS)?;
+                elem_value.set_enum(&[op.input_plug_id as u32]);
+                Ok(true)
+            }
+            _ => Ok(false),
+        }
+    }
+
+    fn write(&self, elem_id: &alsactl::ElemId, _: &alsactl::ElemValue, new: &alsactl::ElemValue)
+        -> Result<bool, Error>
+    {
+        match elem_id.get_name().as_str() {
+            SPDIF_OUT_SRC_NAME => {
+                let mut vals = [0];
+                new.get_enum(&mut vals);
+                let mut op = AudioSelector::new(SPDIF_OUT_SRC_FB_ID, CtlAttr::Current, vals[0] as u8);
+                self.control(&AUDIO_SUBUNIT_0_ADDR, &mut op, FCP_TIMEOUT_MS)?;
+                Ok(true)
+            }
+            _ => Ok(false),
+        }
+    }
+}
+
+impl SpdifOutCtl for BebobAvc {}
