+++ libs/bebob/src/stanton.rs
@@ -12,9 +12,9 @@ use ta1394::{MUSIC_SUBUNIT_0, Ta1394Avc};
 use ta1394::ccm::{SignalAddr, SignalSubunitAddr};
 use ta1394::audio::{AUDIO_SUBUNIT_0_ADDR, AudioFeature, FeatureCtl, CtlAttr, AudioCh};
 
-use crate::bebob::BebobAvc;
-use crate::bebob::common_ctls::ClkCtl;
-use crate::bebob::model::OUT_VOL_NAME;
+use super::BebobAvc;
+use super::common_ctls::ClkCtl;
+use super::model::OUT_VOL_NAME;
 
 pub struct ScratchampModel<'a>{
     avc: BebobAvc,
