--- libs/src/lib.rs
@@ -1,3 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-// Copyright (c) 2020 Takashi Sakamoto
-pub mod bebob;
