+++ services/Cargo.toml
@@ -10,12 +10,12 @@ Service programs to control audio and music unit on IEEE 1394 bus, supported by 
 
 [dependencies]
 libc = "0.2"
-snd_fw_ctl_services = { path = "../libs" }
 dg00x = { path = "../libs/dg00x" }
 tascam = { path = "../libs/tascam" }
 efw = { path = "../libs/efw" }
 motu = { path = "../libs/motu" }
 oxfw = { path = "../libs/oxfw" }
+bebob = { path = "../libs/bebob" }
 
 [[bin]]
 name = "snd-firewire-digi00x-ctl-service"
