+++ services/src/snd-bebob-ctl-service.rs
@@ -1,7 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2020 Takashi Sakamoto
-use snd_fw_ctl_services::bebob;
-
 use std::env;
 
 fn main() {
